defined two goroutine to server the gRPC server at 9000 and 9001 port
